DEV: re-introducing web argument to prevetn accidental new_main erase
