+++ main.go
@@ -13,6 +13,7 @@ import (
 	"fmt"
   "time"
   "os"
+  "flag"
 )
 
 type Weather struct {
@@ -58,7 +59,12 @@ var (
 )
 
 func main() {
-	var err error
+
+    // Parse the "web" flag
+    webFlag := flag.Bool("web", false, "Pass this flag to enable the web server with file check routine")
+    flag.Parse() // Parse command-line flags
+
+  var err error
 
 //	db, err = sql.Open("sqlite3", "./main.db")
 	db, err = sql.Open("sqlite3", "./data/compiled/main.db")
@@ -85,9 +91,11 @@ http.HandleFunc("/privacy-policy", func(w http.ResponseWriter, r *http.Request) 
     http.ServeFile(w, r, "./src/frontend/html/privacy-policy.html")  // Make sure the path is correct
 })
 
-
-	// Start the file checking routine
-	go startFileCheckRoutine()
+// On web server, every 2 hours, check for a new database deilvery, and swap dbs accordingly
+ if *webFlag {
+        // Start the file checking routine
+        go startFileCheckRoutine()
+    }
 
 //liston on all newtowrk interfaces including localhost
 log.Fatal(http.ListenAndServe("0.0.0.0:8080", nil))
