+++ src/frontend/html/index.html
@@ -37,6 +37,7 @@
                 <option value="Berlin" selected>Berlin</option>
                 <option value="Glasgow" >Glasgow</option> <!-- Default selection is Glasgow -->
                 <option value="Edinburgh">Edinburgh</option>
+                <option value="Munich">Munich</option>
             </select>
 </div>
   <div class="form-group">
